requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
firebase-admin==6.4.0
PyPDF2==3.0.1
google-generativeai>=0.4.0
python-dotenv==1.0.1
pydantic==2.6.1
gunicorn==21.2.0 